Implement parsing of flatbuffer response

+++ tests/common/f64_utils.rs
@@ -1,4 +1,4 @@
-pub fn approx_equal(a: f64, b: f64, dp: u8) -> bool {
-    let p = 10f64.powi(-(dp as i32));
+pub fn approx_equal(a: f32, b: f32, dp: u8) -> bool {
+    let p = 10f32.powi(-(dp as i32));
     (a - b).abs() < p
 }
